Fix incorrect kernel signature check
Currently, we found a corner case in our kernel verification where if
you have a different kernel vendor installed on the system (A CentOS
system with a OL kernel), then this verification somestimes may fail,
since we were getting only the first occurrence of the kernel installed
on the system, and not the one that is booted.

This commit is to fix this behavior and get the correct instellad kernel
rather than the first one that it found when querying the installed
kernels on the system.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>